+++ components/Message/MessageController.ts
@@ -1,7 +1,6 @@
-import { Request, Response, NextFunction } from "express";
-import { messageSchema } from "./MessageModel";
-import { userSchema } from "../User/UserModel";
-import { IRequestWithUser } from "../Interfaces";
+import { Request, Response, NextFunction } from 'express';
+import { messageSchema } from './MessageModel';
+import { IRequestWithUser } from '../Interfaces';
 export const getMessages = async (
   req: Request,
   res: Response,
@@ -27,34 +26,36 @@ export const postMessage = async (
       message
     });
     await newMessage.save();
-    return res.status(201).json({ msg: "Message succesfully created" });
+    return res.status(201).json({ msg: 'Message succesfully created' });
   } catch (err) {
     return res.status(400).json({ msg: err });
   }
 };
 export const putMessage = async (
-  req: Request,
+  req: IRequestWithUser,
   res: Response,
   next: NextFunction
 ) => {
-  const { owner, message, _id } = req.body;
+  const { message, _id } = req.body;
+  const user = req.user;
   try {
-    await messageSchema.findByIdAndUpdate({ _id }, { message });
-    res.status(200).json({ msg: "Message updated!" });
+    await messageSchema.findByIdAndUpdate({ _id, owner: user }, { message });
+    res.status(200).json({ msg: 'Message updated!' });
   } catch (err) {
     return res.status(401).json({ msg: err });
   }
 };
 export const deleteMessage = async (
-  req: Request,
+  req: IRequestWithUser,
   res: Response,
   next: NextFunction
 ) => {
   const { _id } = req.body;
+  const user = req.user;
   try {
-    await messageSchema.findOneAndRemove({ _id });
-    return res.status(200).json({ msg: "Sucesfully deleted!" });
+    await messageSchema.findOneAndRemove({ _id, owner: user });
+    return res.status(200).json({ msg: 'Sucesfully deleted!' });
   } catch (err) {
-    return res.status(400).json({ msg: "deleted!!" });
+    return res.status(400).json({ msg: 'deleted!!' });
   }
 };
